rename: Package 이름 오타 수정
