--- src/main/java/com/connect/skilltrade/sample/domain/sub/imfrastructure/SubRepository.java
@@ -1,4 +1,0 @@
-package com.connect.skilltrade.sample.domain.sub.imfrastructure;
-
-public interface SubRepository {
-}
