+++ src/main/java/com/connect/skilltrade/sample/domain/sub/infrastructure/SubReaderImpl.java
@@ -1,4 +1,4 @@
-package com.connect.skilltrade.sample.domain.sub.imfrastructure;
+package com.connect.skilltrade.sample.domain.sub.infrastructure;
 
 import com.connect.skilltrade.sample.domain.sub.domain.SubReader;
 import org.springframework.stereotype.Component;
