+++ src/main/java/com/connect/skilltrade/sample/domain/sub/infrastructure/SubRepository.java
@@ -1,0 +1,4 @@
+package com.connect.skilltrade.sample.domain.sub.infrastructure;
+
+public interface SubRepository {
+}
